chore: Refactor local protocol module: Split into focused submodules
- Created `implementations.rs` for LocalProtocol trait implementations and constructor methods for endpoint types.
- Moved endpoint type definitions from `mod.rs` to `endpoints.rs`, improving modularity and reducing line count.
- Updated `mod.rs` to re-export endpoint types and maintain a clean interface.
- Adjusted tests to reflect new module structure and ensure all tests pass.
- Documented restructuring process and insights in learnings files.
