+++ src/protocol/global/mod.rs
@@ -13,387 +13,19 @@
 //! - **Type Aliases**: Convenience types for common patterns
 //! - **Builder Functions**: Easy construction of complex protocols
 
-use crate::protocol::foundation::{
-    ActionIOTMarker, BiDirectionalAction, ChanId, CommMetadata, DefaultChan, GlobalProtocol,
-    Message, MsgLbl, RequestLbl, ResponseLbl, Role,
+mod protocols;
+mod implementations;
+
+// Re-export all protocol types and make implementations available
+pub use protocols::{
+    TChanSend, TChanRecv, TChanChoice, TChanOffer, 
+    TChanPar, TChanEnd, TChanStart
 };
-use std::fmt::Debug;
-use std::marker::PhantomData;
+// Implementations are included through module system (impl blocks can't be re-exported)
 
-// ============================================================================
-// Core Global Protocol Types
-// ============================================================================
-
-/// Global Type: Represents sending a message over a specific channel
-///
-/// - `S`: Sender Role
-/// - `R`: Receiver Role  
-/// - `C`: Channel ID type
-/// - `L`: Message label type
-/// - `Msg`: Message type being sent
-/// - `P`: Continuation Global Protocol after the send
-/// - `AIO`: ActionIOTMarker specifying required I/O type
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanSend<
-    S: Role,
-    R: Role,
-    C: ChanId,
-    L: MsgLbl,
-    Msg: Message,
-    P: GlobalProtocol,
-    AIO: ActionIOTMarker,
-> {
-    _sender: PhantomData<S>,
-    _receiver: PhantomData<R>,
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<L>,
-    _msg: PhantomData<Msg>,
-    _protocol: PhantomData<P>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<
-        S: Role,
-        R: Role,
-        C: ChanId,
-        L: MsgLbl,
-        Msg: Message,
-        P: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > GlobalProtocol for TChanSend<S, R, C, L, Msg, P, AIO>
-{
-}
-
-impl<
-        S: Role,
-        R: Role,
-        C: ChanId,
-        L: MsgLbl,
-        Msg: Message,
-        P: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > TChanSend<S, R, C, L, Msg, P, AIO>
-{
-    pub fn new() -> Self {
-        Self {
-            _sender: PhantomData,
-            _receiver: PhantomData,
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _msg: PhantomData,
-            _protocol: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-
-    /// Get CommMetadata for this send operation
-    pub fn metadata() -> CommMetadata<C, L>
-    where
-        C: Default,
-        L: Default,
-    {
-        CommMetadata::new(C::default(), L::default())
-    }
-}
-
-/// Global Type: Represents receiving a message over a specific channel
-///
-/// - `R`: Receiver Role
-/// - `S`: Sender Role
-/// - `C`: Channel ID type
-/// - `L`: Message label type  
-/// - `Msg`: Message type being received
-/// - `P`: Continuation Global Protocol after the receive
-/// - `AIO`: ActionIOTMarker specifying required I/O type
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanRecv<
-    R: Role,
-    S: Role,
-    C: ChanId,
-    L: MsgLbl,
-    Msg: Message,
-    P: GlobalProtocol,
-    AIO: ActionIOTMarker,
-> {
-    _receiver: PhantomData<R>,
-    _sender: PhantomData<S>,
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<L>,
-    _msg: PhantomData<Msg>,
-    _protocol: PhantomData<P>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<
-        R: Role,
-        S: Role,
-        C: ChanId,
-        L: MsgLbl,
-        Msg: Message,
-        P: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > GlobalProtocol for TChanRecv<R, S, C, L, Msg, P, AIO>
-{
-}
-
-impl<
-        R: Role,
-        S: Role,
-        C: ChanId,
-        L: MsgLbl,
-        Msg: Message,
-        P: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > TChanRecv<R, S, C, L, Msg, P, AIO>
-{
-    pub fn new() -> Self {
-        Self {
-            _receiver: PhantomData,
-            _sender: PhantomData,
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _msg: PhantomData,
-            _protocol: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-
-    /// Get CommMetadata for this receive operation
-    pub fn metadata() -> CommMetadata<C, L>
-    where
-        C: Default,
-        L: Default,
-    {
-        CommMetadata::new(C::default(), L::default())
-    }
-}
-
-/// Global Type: Represents a choice between protocol branches
-///
-/// - `R`: Role making the choice
-/// - `C`: Channel ID type
-/// - `Lbl`: Message label type for the choice point
-/// - `Left`: Left branch protocol
-/// - `Right`: Right branch protocol  
-/// - `AIO`: ActionIOTMarker for choice communication
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanChoice<
-    R: Role,
-    C: ChanId,
-    Lbl: MsgLbl,
-    Left: GlobalProtocol,
-    Right: GlobalProtocol,
-    AIO: ActionIOTMarker,
-> {
-    _chooser: PhantomData<R>,
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<Lbl>,
-    _left: PhantomData<Left>,
-    _right: PhantomData<Right>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<
-        R: Role,
-        C: ChanId,
-        Lbl: MsgLbl,
-        Left: GlobalProtocol,
-        Right: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > GlobalProtocol for TChanChoice<R, C, Lbl, Left, Right, AIO>
-{
-}
-
-impl<
-        R: Role,
-        C: ChanId,
-        Lbl: MsgLbl,
-        Left: GlobalProtocol,
-        Right: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > TChanChoice<R, C, Lbl, Left, Right, AIO>
-{
-    pub fn new() -> Self {
-        Self {
-            _chooser: PhantomData,
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _left: PhantomData,
-            _right: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-}
-
-/// Global Type: Represents offering branches to another role
-///
-/// - `R`: Role offering the branches (dual to chooser)
-/// - `C`: Channel ID type
-/// - `Lbl`: Message label type for the offer point
-/// - `Left`: Left branch protocol
-/// - `Right`: Right branch protocol  
-/// - `AIO`: ActionIOTMarker for offer communication
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanOffer<
-    R: Role,
-    C: ChanId,
-    Lbl: MsgLbl,
-    Left: GlobalProtocol,
-    Right: GlobalProtocol,
-    AIO: ActionIOTMarker,
-> {
-    _offerer: PhantomData<R>,
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<Lbl>,
-    _left: PhantomData<Left>,
-    _right: PhantomData<Right>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<
-        R: Role,
-        C: ChanId,
-        Lbl: MsgLbl,
-        Left: GlobalProtocol,
-        Right: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > GlobalProtocol for TChanOffer<R, C, Lbl, Left, Right, AIO>
-{
-}
-
-impl<
-        R: Role,
-        C: ChanId,
-        Lbl: MsgLbl,
-        Left: GlobalProtocol,
-        Right: GlobalProtocol,
-        AIO: ActionIOTMarker,
-    > TChanOffer<R, C, Lbl, Left, Right, AIO>
-{
-    pub fn new() -> Self {
-        Self {
-            _offerer: PhantomData,
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _left: PhantomData,
-            _right: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-}
-
-/// Global Type: Represents parallel composition of protocols
-///
-/// - `C`: Channel ID type
-/// - `Lbl`: Message label type for parallel execution context
-/// - `Left`: Left parallel branch
-/// - `Right`: Right parallel branch
-/// - `IsDisjoint`: Marker ensuring branches are disjoint (must be Send + Sync + Debug)
-/// - `AIO`: ActionIOTMarker for parallel coordination
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanPar<
-    C: ChanId,
-    Lbl: MsgLbl,
-    Left: GlobalProtocol,
-    Right: GlobalProtocol,
-    IsDisjoint: Send + Sync + 'static + Debug,
-    AIO: ActionIOTMarker,
-> {
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<Lbl>,
-    _left: PhantomData<Left>,
-    _right: PhantomData<Right>,
-    _disjoint: PhantomData<IsDisjoint>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<
-        C: ChanId,
-        Lbl: MsgLbl,
-        Left: GlobalProtocol,
-        Right: GlobalProtocol,
-        IsDisjoint: Send + Sync + 'static + Debug,
-        AIO: ActionIOTMarker,
-    > GlobalProtocol for TChanPar<C, Lbl, Left, Right, IsDisjoint, AIO>
-{
-}
-
-impl<
-        C: ChanId,
-        Lbl: MsgLbl,
-        Left: GlobalProtocol,
-        Right: GlobalProtocol,
-        IsDisjoint: Send + Sync + 'static + Debug,
-        AIO: ActionIOTMarker,
-    > TChanPar<C, Lbl, Left, Right, IsDisjoint, AIO>
-{
-    pub fn new() -> Self {
-        Self {
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _left: PhantomData,
-            _right: PhantomData,
-            _disjoint: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-}
-
-/// Global Type: Represents protocol termination
-///
-/// - `C`: Channel ID type
-/// - `L`: Message label type for termination context
-/// - `AIO`: ActionIOTMarker for cleanup operations
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanEnd<C: ChanId, L: MsgLbl, AIO: ActionIOTMarker> {
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<L>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<C: ChanId, L: MsgLbl, AIO: ActionIOTMarker> GlobalProtocol for TChanEnd<C, L, AIO> {}
-
-impl<C: ChanId, L: MsgLbl, AIO: ActionIOTMarker> TChanEnd<C, L, AIO> {
-    pub fn new() -> Self {
-        Self {
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-}
-
-/// Global Type: Represents protocol initialization
-///
-/// - `C`: Channel ID type
-/// - `L`: Message label type for initialization context
-/// - `Start`: Continuation protocol after start
-/// - `AIO`: ActionIOTMarker for initialization operations
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TChanStart<C: ChanId, L: MsgLbl, Start: GlobalProtocol, AIO: ActionIOTMarker> {
-    _chan: PhantomData<C>,
-    _lbl: PhantomData<L>,
-    _start: PhantomData<Start>,
-    _aio: PhantomData<AIO>,
-}
-
-impl<C: ChanId, L: MsgLbl, Start: GlobalProtocol, AIO: ActionIOTMarker> GlobalProtocol
-    for TChanStart<C, L, Start, AIO>
-{
-}
-
-impl<C: ChanId, L: MsgLbl, Start: GlobalProtocol, AIO: ActionIOTMarker>
-    TChanStart<C, L, Start, AIO>
-{
-    pub fn new() -> Self {
-        Self {
-            _chan: PhantomData,
-            _lbl: PhantomData,
-            _start: PhantomData,
-            _aio: PhantomData,
-        }
-    }
-}
+use crate::protocol::foundation::{
+    BiDirectionalAction, DefaultChan, RequestLbl, ResponseLbl,
+};
 
 // ============================================================================
 // Type Aliases for Common Patterns
