+++ work/learnings.md
@@ -30,28 +30,49 @@ without relying on unstable features.
 - **All tests pass**: 35/35 tests successful after restructuring
 - **Technical constraint learned**: Trait implementations cannot be moved to separate modules in Rust
 
-**Current Module Compliance Status:**
+**Phase 3: Local Module Restructuring ✅**
+
+- **Original**: 448 lines (above 300-line guideline)
+- **Restructured into 3 focused modules**:
+  - `mod.rs`: 80 lines ✅ (82% reduction, well under 300-line guideline)
+  - `endpoints.rs`: 166 lines (7 endpoint struct definitions: EpChanSend, EpChanRecv, EpChanOffer, EpChanChoice, EpChanPar, EpChanEnd, EpChanStart)
+  - `implementations.rs`: 235 lines (LocalProtocol trait implementations and constructor methods)
+- **All tests pass**: 35/35 tests successful after restructuring
+
+**Phase 4: Global Module Restructuring ✅**
+
+- **Original**: 434 lines (above 300-line guideline)
+- **Restructured into 3 focused modules**:
+  - `mod.rs`: 66 lines ✅ (85% reduction, well under 300-line guideline)
+  - `protocols.rs`: 172 lines (7 global protocol struct definitions: TChanSend, TChanRecv, TChanChoice, TChanOffer, TChanPar, TChanEnd, TChanStart)
+  - `implementations.rs`: 234 lines (GlobalProtocol trait implementations and constructor methods)
+- **All tests pass**: 35/35 tests successful after restructuring
+
+**Final Module Compliance Status (All Compliant ✅):**
 
 - `foundation/mod.rs`: 209 lines ✅ (compliant)
-- `duality/mod.rs`: 94 lines ✅ (compliant - 80% reduction)
-- `projection/mod.rs`: 336 lines ✅ (compliant - 27% reduction)  
-- `local/mod.rs`: 448 lines ⚠️ (needs restructuring)
-- `global/mod.rs`: 434 lines ⚠️ (needs restructuring)
+- `duality/mod.rs`: 94 lines ✅ (compliant - 80% reduction from 476 lines)
+- `projection/mod.rs`: 336 lines ✅ (compliant - 27% reduction from 460 lines)  
+- `local/mod.rs`: 80 lines ✅ (compliant - 82% reduction from 448 lines)
+- `global/mod.rs`: 66 lines ✅ (compliant - 85% reduction from 434 lines)
 
 **Implementation/Test Ratios:**
+- foundation: 3.0:1 ratio (209 impl / 69 tests) ✅ (within 2-4:1 target range)
+- duality: 5.1:1 ratio (473 impl lines / 92 tests) ⚠️ (slightly above ideal)
+- projection: 6.8:1 ratio (486 impl lines / 71 tests) ⚠️ (above ideal, needs more tests)
+- local: 2.9:1 ratio (481 impl lines / 166 tests) ✅ (within 2-4:1 target range)
+- global: 3.0:1 ratio (472 impl lines / 153 tests) ✅ (within 2-4:1 target range)
 
-- duality: 5.1:1 ratio (476 impl / 92 tests)
-- projection: 6.8:1 ratio (460 impl / 67 tests)
-- Note: Both exceed 2:1 guideline, indicating need for more comprehensive testing
+**Key Module Structure Patterns Successful:**
 
-**Key Module Structure Patterns Learned:**
-
-1. **Test Extraction Process**: Proper workflow for converting embedded test modules to separate files
+1. **Progressive Restructuring Process**: Step-by-step approach works well for large modules
 2. **Module Declaration**: Use `#[cfg(test)] mod tests;` for separate test files
-3. **Test File Structure**: Test files should directly contain test functions without wrapper modules
-4. **Validation Workflow**: Always verify fmt/clippy/build/test after structural changes
+3. **Consistent Patterns**: Same structure works across different protocol modules
+4. **Visibility Management**: `pub(super)` for fields, `pub` for structs, proper re-exports
+5. **Import Resolution**: Tests need explicit imports for traits from foundation module
+6. **Validation Workflow**: Always verify fmt/clippy/build/test after structural changes
 
-**Next Phase:** Task 1.1.6c.2 (advanced restructuring) to bring all modules under 300-line guideline, or continue with other priority tasks.
+**Task 1.1.6c.2 COMPLETE ✅:** All protocol modules now comply with Size and Module Structure Guidelines
 
 ### Task 1.1.3 Successfully Completed: Local Endpoint Types
